requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
opencv-python==4.8.1.78
pyzbar==0.1.9
pyserial==3.5
requests==2.31.0
